recommendation/requirements.txt
flask==2.0.1
Werkzeug==2.0.1
numpy
pandas
scikit-surprise==1.1.3